Перешел на int
Добавил в тесты проекцию лимитов

Изменил аннотации к ассертам

+++ internal/transport/nats/events.go
@@ -52,35 +52,35 @@ type WalletCreatedPayload struct {
 	WalletType      WalletType   `json:"wallet_type"`
 	WalletStatus    WalletStatus `json:"wallet_status"`
 	Balance         string       `json:"balance"`
-	CreatedAt       int64        `json:"created_at"`
-	UpdatedAt       int64        `json:"updated_at"`
+	CreatedAt       int          `json:"created_at"`
+	UpdatedAt       int          `json:"updated_at"`
 	IsDefault       bool         `json:"is_default"`
 	IsBasic         bool         `json:"is_basic"`
 }
 
 type SetDefaultStartedPayload struct {
 	UUID      string `json:"uuid"`
-	CreatedAt int64  `json:"created_at"`
+	CreatedAt int    `json:"created_at"`
 }
 
 type DefaultUnsettedPayload struct {
 	UUID      string `json:"uuid"`
-	CreatedAt int64  `json:"created_at"`
+	CreatedAt int    `json:"created_at"`
 }
 
 type DefaultSettedPayload struct {
 	UUID      string `json:"uuid"`
-	CreatedAt int64  `json:"created_at"`
+	CreatedAt int    `json:"created_at"`
 }
 
 type WalletDisabledPayload struct {
-	CreatedAt int64 `json:"created_at"`
+	CreatedAt int `json:"created_at"`
 }
 
 type BlockersSettedPayload struct {
-	IsGamblingActive bool  `json:"is_gambling_active"`
-	IsBettingActive  bool  `json:"is_betting_active"`
-	CreatedAt        int64 `json:"created_at"`
+	IsGamblingActive bool `json:"is_gambling_active"`
+	IsBettingActive  bool `json:"is_betting_active"`
+	CreatedAt        int  `json:"created_at"`
 }
 
 type LimitChangedV2 struct {
@@ -91,8 +91,8 @@ type LimitChangedV2 struct {
 		IntervalType string `json:"interval_type"`
 		Amount       string `json:"amount"`
 		CurrencyCode string `json:"currency_code"`
-		StartedAt    int64  `json:"started_at"`
-		ExpiresAt    int64  `json:"expires_at"`
+		StartedAt    int    `json:"started_at"`
+		ExpiresAt    int    `json:"expires_at"`
 		Status       bool   `json:"status"`
 	} `json:"limits"`
 }
